requirements.txt
Flask
pandas
numpy
spotipy
scikit-learn
python-dotenv
joblib
scipy
